+++ Tests.Integration.Presentation.Web/Tools/ItSystemHelper.cs
@@ -241,66 +241,5 @@ namespace Tests.Integration.Presentation.Web.Tools
 
             return await HttpApi.GetWithCookieAsync(url, cookie);
         }
-
-        public static async Task<HttpResponseMessage> SendPostRelationAsync(CreateSystemRelationDTO input, Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl("api/v1/systemrelations");
-
-            return await HttpApi.PostWithCookieAsync(url, login, input);
-        }
-
-        public static async Task<HttpResponseMessage> SendDeleteRelationAsync(int systemUsageId, int relationId, Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/from/{systemUsageId}/{relationId}");
-
-            return await HttpApi.DeleteWithCookieAsync(url, login);
-        }
-
-        public static async Task<IEnumerable<SystemRelationDTO>> GetRelationsAsync(int systemUsageId, Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/from/{systemUsageId}");
-
-            using (var response = await HttpApi.GetWithCookieAsync(url, login))
-            {
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                return await response.ReadResponseBodyAsKitosApiResponseAsync<IEnumerable<SystemRelationDTO>>();
-            }
-        }
-
-        public static async Task<IEnumerable<NamedEntityDTO>> GetAvailableDestinationSystemsAsync(int systemUsageId, string prefix,Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/options/{systemUsageId}/available-destination-systems?nameContent={prefix}&amount=25");
-
-            using (var response = await HttpApi.GetWithCookieAsync(url, login))
-            {
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                return await response.ReadResponseBodyAsKitosApiResponseAsync<IEnumerable<NamedEntityDTO>>();
-            }
-        }
-
-        public static async Task<SystemRelationOptionsDTO> GetAvailableOptionsAsync(int systemUsageId, int targetUsageId, Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/options/{systemUsageId}/in-relation-to/{targetUsageId}");
-
-            using (var response = await HttpApi.GetWithCookieAsync(url, login))
-            {
-                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-                return await response.ReadResponseBodyAsKitosApiResponseAsync<SystemRelationOptionsDTO>();
-            }
-        }
-
-        public static async Task<HttpResponseMessage> SendGetRelationAsync(int systemUsageId, int relationId, Cookie login = null)
-        {
-            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
-            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/from/{systemUsageId}/{relationId}");
-
-            return await HttpApi.GetWithCookieAsync(url, login);
-        }
-
     }
 }
