+++ Tests.Integration.Presentation.Web/Tools/SystemRelationHelper.cs
@@ -3,6 +3,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Core.DomainModel.Organization;
+using Presentation.Web.Models;
 using Presentation.Web.Models.SystemRelations;
 using Xunit;
 
@@ -10,11 +11,6 @@ namespace Tests.Integration.Presentation.Web.Tools
 {
     public class SystemRelationHelper
     {
-        public static SystemRelationDTO CreateSystemRelationDTO()
-        {
-            return new SystemRelationDTO();
-        }
-
         public static async Task<IEnumerable<SystemRelationDTO>> SendGetRelationsAsync(int systemUsageId, Cookie login = null)
         {
             login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
@@ -57,6 +53,43 @@ namespace Tests.Integration.Presentation.Web.Tools
             var url = TestEnvironment.CreateUrl("api/v1/systemrelations");
 
             return await HttpApi.PatchWithCookieAsync(url, login, input);
+        }
+
+
+        public static async Task<IEnumerable<SystemRelationDTO>> GetRelationsAsync(int systemUsageId, Cookie login = null)
+        {
+            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
+            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/from/{systemUsageId}");
+
+            using (var response = await HttpApi.GetWithCookieAsync(url, login))
+            {
+                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+                return await response.ReadResponseBodyAsKitosApiResponseAsync<IEnumerable<SystemRelationDTO>>();
+            }
+        }
+
+        public static async Task<IEnumerable<NamedEntityDTO>> GetAvailableDestinationSystemsAsync(int systemUsageId, string prefix, Cookie login = null)
+        {
+            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
+            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/options/{systemUsageId}/available-destination-systems?nameContent={prefix}&amount=25");
+
+            using (var response = await HttpApi.GetWithCookieAsync(url, login))
+            {
+                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+                return await response.ReadResponseBodyAsKitosApiResponseAsync<IEnumerable<NamedEntityDTO>>();
+            }
+        }
+
+        public static async Task<SystemRelationOptionsDTO> GetAvailableOptionsAsync(int systemUsageId, int targetUsageId, Cookie login = null)
+        {
+            login = login ?? await HttpApi.GetCookieAsync(OrganizationRole.GlobalAdmin);
+            var url = TestEnvironment.CreateUrl($"api/v1/systemrelations/options/{systemUsageId}/in-relation-to/{targetUsageId}");
+
+            using (var response = await HttpApi.GetWithCookieAsync(url, login))
+            {
+                Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+                return await response.ReadResponseBodyAsKitosApiResponseAsync<SystemRelationOptionsDTO>();
+            }
         }
     }
 }
